backend/ai-module/requirements.txt
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
cloudpickle==3.1.0
colorama==0.4.6
contourpy==1.3.0
cycler==0.12.1
dnspython==2.7.0
flask==3.0.3
Flask-Cors==5.0.0
fonttools==4.54.1
gunicorn==23.0.0
idna==3.10
importlib-metadata==8.5.0
importlib-resources==6.4.5
itsdangerous==2.2.0
jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.7
llvmlite==0.43.0
MarkupSafe==3.0.1
matplotlib==3.7.2
mongoengine==0.29.1
numba==0.60.0
numpy==1.26.1
packaging==24.1
pandas==2.1.2
pillow==11.0.0
psycopg2-binary==2.9.10
pymongo==4.10.1
pyparsing==3.0.9
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
scikit-learn==1.5.2
scipy==1.13.1
seaborn==0.12.2
shap==0.46.0
six==1.16.0
slicer==0.0.8
threadpoolctl==3.5.0
tqdm==4.66.5
tzdata==2024.2
urllib3==2.2.3
werkzeug==3.0.4
xgboost==2.1.1
zipp==3.20.2